feat : 로또 통계를 구해줄 컴퓨터 클래스 구현 및 테스트 추가
