+++ src/test/java/lotto/ComputerTest.java
@@ -1,13 +1,11 @@
-package lotto.service;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+package lotto;
 
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.stream.IntStream;
-import lotto.Computer;
-import lotto.Lotto;
+import lotto.service.ComputerSpy;
+import lotto.service.LottoService;
 import lotto.validator.Validator;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -15,47 +13,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class LottoServiceTest {
-
+public class ComputerTest {
     private static Collection<Arguments> param1() {
-        return Arrays.asList(
-            Arguments.of("1000원으로 1개 구매", "1000", 1),
-            Arguments.of("9000원으로 9개 구매", "9000", 9),
-            Arguments.of("10000원으로 10개 구매", "10000", 10),
-            Arguments.of("1000000원으로 1000개 구매", "1000000", 1000)
-        );
-    }
-
-    @ParameterizedTest(name = "{index}: {0}")
-    @MethodSource("param1")
-    @DisplayName("정상적인 금액이 입력되면 전부 로또를 구매 해야한다.")
-    void test1(String description, String money, int size) {
-        LottoService service = new LottoService(new Validator(), new Computer());
-        List<Lotto> lottos = service.buy(money);
-        Assertions.assertEquals(size, lottos.size());
-    }
-
-    private static Collection<Arguments> param2() {
-        return Arrays.asList(
-            Arguments.of("1000j 비정상적인 입력테스트", "1000j"),
-            Arguments.of("abcd 비정상적인 입력테스트", "abcd"),
-            Arguments.of("-1000 비정상적인 입력테스트", "-1000"),
-            Arguments.of("999 비정상적인 입력테스트", "999"),
-            Arguments.of("1 비정상적인 입력테스트", "1")
-        );
-    }
-
-    @ParameterizedTest(name = "{index}: {0}")
-    @MethodSource("param2")
-    @DisplayName("비정상적인 금액이 입력되면 익셉션을 던진다")
-    void test2(String description, String money) {
-        LottoService service = new LottoService(new Validator(), new Computer());
-        IllegalArgumentException e = Assertions.assertThrows(IllegalArgumentException.class,
-            () -> service.buy(money));
-        assertThat(e.getMessage()).contains("[ERROR]");
-    }
-
-    private static Collection<Arguments> param3() {
         return Arrays.asList(
             Arguments.of("1등이 한번 등장",
                 Arrays.asList(new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6))),
@@ -109,14 +68,13 @@ class LottoServiceTest {
     }
 
     @ParameterizedTest(name = "{index}: {0}")
-    @MethodSource("param3")
-    @DisplayName("로또 서비스의 getResult 테스트")
-    void test3(String description, List<Lotto> lottos, List<Integer> win, int bonus,
+    @MethodSource("param1")
+    @DisplayName("구매한 로또 목록, 당첨 번호, 보너스 번호를 받아 결과를 리턴하는지 테스트")
+    void test1(String description, List<Lotto> lottos, List<Integer> win, int bonus,
         List<Integer> result) {
-        ComputerSpy computerSpy = new ComputerSpy();
-        LottoService service = new LottoService(new Validator(), computerSpy);
-        List<Integer> a = service.getResult(lottos, win, bonus);
-        Assertions.assertTrue(computerSpy.isCalled);
+        Computer computer = new Computer();
+        computer.init();
+        List<Integer> a = computer.getResult(lottos, win, bonus);
+        IntStream.range(0, 5).forEach(i -> Assertions.assertEquals(result.get(i), a.get(i)));
     }
-
 }
